Refactoring setWebLifecycleState
